+++ plone/autoform/testing.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
-from plone.testing import Layer
 from plone.testing import z2
 from plone.testing import zca
+
 import plone.autoform
 
 
@@ -9,8 +9,8 @@ AUTOFORM_FIXTURE = zca.ZCMLSandbox(
     bases=(z2.STARTUP,),
     filename='configure.zcml',
     package=plone.autoform,
-    name="plone.autoform:Fixture")
+    name='plone.autoform:Fixture')
 
 AUTOFORM_INTEGRATION_TESTING = z2.IntegrationTesting(
     bases=(AUTOFORM_FIXTURE,),
-    name="plone.autoform:Integration")
+    name='plone.autoform:Integration')