+++ plone/autoform/tests/test_widgets.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from plone.testing.zca import UNIT_TESTING
+
 import unittest2 as unittest
 
 
@@ -16,6 +17,7 @@ class TestParameterizedWidget(unittest.TestCase):
 
         @implementer(IWidget)
         class DummyWidget(object):
+
             def __init__(self, request):
                 self.request = request
 
@@ -39,6 +41,7 @@ class TestParameterizedWidget(unittest.TestCase):
         from zope.schema import Field
 
         class DummyWidget(object):
+
             def __init__(self, request):
                 self.request = request